Added constant tests

+++ Mathema.Models/Constants/Constant.cs
@@ -9,12 +9,12 @@ namespace Mathema.Models.Constants
     {
         public string Symbol { get; }
 
-        public double Value { get; }
+        public decimal Value { get; }
 
         public ConstantTypes Type { get; }
 
 
-        public Constant(string symbol, double value, ConstantTypes type)
+        public Constant(string symbol, decimal value, ConstantTypes type)
         {
             this.Symbol = symbol;
             this.Value = value;
