+++ Mathema.Shared/Constants/Constants.cs
@@ -4,12 +4,12 @@ namespace Mathema.Shared.Constants
 {
     public class Constants
     {
-        public static double PI { get; } = Math.PI;
+        public static decimal PI { get; } = (decimal)Math.PI;
 
-        public static double e { get; } = Math.E;
+        public static decimal e { get; } = (decimal)Math.E;
 
-        public static double EM { get; } = 0.577215664901532860606512090082402431042159335;
+        public static decimal EM { get; } = 0.577215664901532860606512090082402431042159335m;
 
-        public static double Phi { get; } = (1.0 * Math.Sqrt(5)) / 2.0;
+        public static decimal Phi { get; } = (decimal)((1.0 * Math.Sqrt(5)) / 2.0);
     }
 }
